requirements.txt
# Google API dependencies
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.88.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0

# HTTP requests (if using raw HTTP approach)
requests>=2.28.0

# Additional utilities
asyncio-compat>=0.1.2